requirements.txt
streamlit==1.28.0
pandas==2.1.0
pillow==10.0.0
pytesseract==0.3.10
easyocr==1.7.0
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
numpy==1.25.2